move nft from eth to coin crate
