+++ mm2src/coins/nft/nft_errors.rs
