+++ mm2src/coins/nft/nft_structs.rs
